feat(sdk)!: move ICA helpers to subpath export
