+++ .changeset/sdk-subpath-exports.md
@@ -1,0 +1,5 @@
+---
+'@hyperlane-xyz/sdk': major
+---
+
+Move ICA call helper exports from the SDK root and `middleware/account/InterchainAccount` to `middleware/account/icaCalls`, and expose fee, middleware account, and utility modules through package subpath exports.
