+++ typescript/sdk/package.json
@@ -25,6 +25,10 @@
       "types": "./dist/core/*.d.ts",
       "default": "./dist/core/*.js"
     },
+    "./fee/*": {
+      "types": "./dist/fee/*.d.ts",
+      "default": "./dist/fee/*.js"
+    },
     "./hook/*": {
       "types": "./dist/hook/*.d.ts",
       "default": "./dist/hook/*.js"
@@ -36,6 +40,10 @@
     "./metadata/*": {
       "types": "./dist/metadata/*.d.ts",
       "default": "./dist/metadata/*.js"
+    },
+    "./middleware/account/*": {
+      "types": "./dist/middleware/account/*.d.ts",
+      "default": "./dist/middleware/account/*.js"
     },
     "./providers/builders/*": {
       "types": "./dist/providers/builders/*.d.ts",
@@ -112,6 +120,10 @@
     "./types": {
       "types": "./dist/types.d.ts",
       "default": "./dist/types.js"
+    },
+    "./utils/*": {
+      "types": "./dist/utils/*.d.ts",
+      "default": "./dist/utils/*.js"
     },
     "./warp/*": {
       "types": "./dist/warp/*.d.ts",
